Separete keyboards into partal classes
